# nextjs-client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BackOffice.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Tenant": {
      "get": {
        "tags": [
          "Tenant"
        ],
        "operationId": "GetTenantsByUserId",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Tenant"
        ],
        "operationId": "SetTenantId",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TenantRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TenantRequestDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Enter the Bearer Authorization string as following: `Bearer Generated-JWT-Token`",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}